+++ CHANGELOG.md
@@ -8,6 +8,18 @@ This changelog follows the [Keep a Changelog](https://keepachangelog.com) format
 ---
 
 
+## [Unreleased]
+
+### Added ###
+- [[#16]] Strict mode to configure warning and error behavior.
+
+### Security ###
+- Update dependencies to remove vulnerabilities.
+
+
+---
+
+
 ## [5.4.0] - 2026-03-26
 
 ### Added ###